Simplify source/provider structure and decompose UI code
- Remove dead code and unused members across components, providers, and installers
- Consolidate duplicated install/update logic (background-install confirmation,
  install queue, pending-update classification, notification IDs)
- Split source_provider.dart into models/, services/, and app_sources/app_source.dart;
  make app JSON migration a standalone library with appFromStoredJson()
- Decompose large page builds and generated-form internals; drop double rebuilds
  and the string-split sub-form key hack
- Add an O(1) source registry with cached host matching and ensureHttpSuccess()
- Fix the dead two-pane back handler so BACK clears the selected app
