+++ lib/app_sources/huaweiappgallery.dart
@@ -205,9 +205,7 @@ class HuaweiAppGallery extends AppSource {
         .join('&');
     final apiUrl = 'https://$host$_apiPath';
     final res = await sourceRequest(apiUrl, mergedSettings, postBody: body);
-    if (res.statusCode != 200) {
-      throw getObtainiumHttpError(res);
-    }
+    ensureHttpSuccess(res);
     final decoded = jsonDecode(utf8.decode(res.bodyBytes));
     if (decoded is! Map<String, dynamic>) {
       throw ObtainiumError(tr('unexpectedStoreApiResponse'), unexpected: true);
