+++ lib/utils/string_utils.dart
@@ -7,3 +7,9 @@ bool isObtainiumVariant(String id) =>
     id == obtainiumId ||
     id == '$obtainiumId.fdroid' ||
     id == '$obtainiumId.debug';
+
+/// Builds a regex alternation pattern from a list of hostname strings,
+/// escaping dots.
+String getSourceRegex(List<String> hosts) {
+  return '(${hosts.join('|').replaceAll('.', '\\.')})';
+}
