+++ lib/utils/url_utils.dart
@@ -1,0 +1,41 @@
+import 'package:obtainium/custom_errors.dart';
+
+/// Ensures the URL is well-formed and starts with HTTPS.
+String preStandardizeUrl(String url) {
+  final firstDotIndex = url.indexOf('.');
+  if (!(firstDotIndex >= 0 && firstDotIndex != url.length - 1) &&
+      !url.contains('[')) {
+    throw UnsupportedURLError();
+  }
+  if (!url.toLowerCase().startsWith('http://') &&
+      !url.toLowerCase().startsWith('https://')) {
+    url = 'https://$url';
+  }
+  final uri = Uri.tryParse(url);
+  final trailingSlash =
+      ((uri?.path.endsWith('/') ?? false) ||
+          ((uri?.path.isEmpty ?? false) && url.endsWith('/'))) &&
+      (uri?.queryParameters.isEmpty ?? false);
+
+  // Only normalize duplicate slashes in the scheme/host/path portion; leave the
+  // query string and fragment untouched so any slashes they contain (e.g. a URL
+  // passed as a query parameter) aren't mangled.
+  var splitIndex = url.length;
+  final queryStart = url.indexOf('?');
+  if (queryStart >= 0 && queryStart < splitIndex) {
+    splitIndex = queryStart;
+  }
+  final fragmentStart = url.indexOf('#');
+  if (fragmentStart >= 0 && fragmentStart < splitIndex) {
+    splitIndex = fragmentStart;
+  }
+  var mainPart = url.substring(0, splitIndex);
+  final rest = url.substring(splitIndex);
+  mainPart = mainPart
+      .split('/')
+      .where((e) => e.isNotEmpty)
+      .join('/')
+      .replaceFirst(':/', '://');
+  url = mainPart + (trailingSlash ? '/' : '') + rest;
+  return url;
+}
